+++ packages/article/src/article.js
@@ -48,7 +48,7 @@ const Article = props => {
 
   return (
     <Responsive>
-      <MessageManager animate delay={3000} scale={scales.medium}>
+      <MessageManager delay={3000} scale={scales.medium}>
         <Component {...newProps} onImagePress={onImagePressArticle} />
       </MessageManager>
     </Responsive>