+++ packages/topic/__tests__/shared-with-style.js
@@ -1,4 +1,3 @@
-import { AppRegistry } from "react-native-web";
 import {
   addSerializers,
   compose,
@@ -6,33 +5,10 @@ import {
   hoistStyleTransform,
   minimaliseTransform,
   minimalWebTransform,
-  rnwTransform,
   stylePrinter
 } from "@times-components/jest-serializer";
 import shared from "./shared-with-style.base";
 import topic from "./fixtures";
-
-const styles = [
-  "alignItems",
-  "backgroundColor",
-  "borderBottomColor",
-  "borderBottomWidth",
-  "borderStyle",
-  "borderTopColor",
-  "borderTopWidth",
-  "color",
-  "flexDirection",
-  "fontFamily",
-  "fontSize",
-  "justifyContent",
-  "lineHeight",
-  "marginBottom",
-  "minHeight",
-  "paddingHorizontal",
-  "paddingTop",
-  "textAlign",
-  "width"
-];
 
 export default () => {
   addSerializers(
@@ -44,8 +20,7 @@ export default () => {
         (value, key) => key !== "style" && key !== "className"
       ),
       flattenStyleTransform,
-      hoistStyleTransform,
-      rnwTransform(AppRegistry, styles)
+      hoistStyleTransform
     )
   );