+++ packages/tracking/__tests__/list-component.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
-import { Text, FlatList } from "react-native";
+import { Text } from "react-native";
+import { TcFlatList } from "@times-components/utils";
 import PropTypes from "prop-types";
 
 class ListComponent extends Component {
@@ -42,16 +43,10 @@ class ListComponent extends Component {
   render() {
     const { items } = this.props;
     return (
-      <FlatList
+      <TcFlatList
         data={items}
         initialNumToRender={items.length}
-        keyExtractor={({ someKey }) => someKey}
-        onViewableItemsChanged={this.onViewableItemsChanged}
         renderItem={({ item }) => <Text>Item {item.someValue}</Text>}
-        viewabilityConfig={{
-          viewAreaCoveragePercentThreshold: 100,
-          waitForInteraction: false
-        }}
       />
     );
   }