+++ packages/ts-components/src/components/recommended-articles/RecommendedArticles.stories.tsx
@@ -33,16 +33,16 @@ storiesOf('Typescript Component/Recommended Articles', module)
   ))
   .add('Recommended Articles - 1 Article', () => (
     <FetchProvider previewData={getArticles(previewData, 1)}>
-      <RecommendedArticles heading="Today's news" isVisible />
+      <RecommendedArticles heading="Today's news" />
     </FetchProvider>
   ))
   .add('Recommended Articles - 2 Article', () => (
     <FetchProvider previewData={getArticles(previewData, 2)}>
-      <RecommendedArticles heading="Today's business" isVisible />
+      <RecommendedArticles heading="Today's business" />
     </FetchProvider>
   ))
   .add('Recommended Articles - 3 Article', () => (
     <FetchProvider previewData={previewData}>
-      <RecommendedArticles heading="Today's sport" isVisible />
+      <RecommendedArticles heading="Today's sport" />
     </FetchProvider>
   ));