+++ packages/utils/src/index.js
@@ -18,6 +18,7 @@ export { default as ServerClientRender } from "./server-client-render";
 export { default as appendToImageURL } from "./append-to-image-url";
 export { default as getSectionFromTiles } from "./get-section-from-tiles";
 export { default as TcView } from "./tc-view";
+export { default as TcFlatList } from "./tc-flat-list";
 export { default as TcText } from "./tc-text";
 export { default as TcTextLink } from "./tc-text-link";
 export { default as TcScrollView } from "./tc-scroll-view";