+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.19.5](https://github.com/newsuk/times-components/compare/@times-components/ad@2.19.4...@times-components/ad@2.19.5) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/ad
+
+
+
+
+
+## [2.19.4](https://github.com/newsuk/times-components/compare/@times-components/ad@2.19.3...@times-components/ad@2.19.4) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/ad
+
+
+
+
+
+## [2.19.3](https://github.com/newsuk/times-components/compare/@times-components/ad@2.19.2...@times-components/ad@2.19.3) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/ad
+
+
+
+
+
+## [2.19.2](https://github.com/newsuk/times-components/compare/@times-components/ad@2.19.1...@times-components/ad@2.19.2) (2022-07-11)
+
+
+### Bug Fixes
+
+* **TDP-1704:** Remove react-native from jest-serializer package ([#3015](https://github.com/newsuk/times-components/issues/3015)) ([5b6da8f](https://github.com/newsuk/times-components/commit/5b6da8f1392c3b42a9c0944d44be3d9ed3e7db2f))
+
+
+
+
+
 ## [2.19.1](https://github.com/newsuk/times-components/compare/@times-components/ad@2.19.0...@times-components/ad@2.19.1) (2022-07-01)