+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.25](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.24...@times-components/article-extras@0.15.25) (2022-07-13)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.15.24](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.23...@times-components/article-extras@0.15.24) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.15.23](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.22...@times-components/article-extras@0.15.23) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.15.22](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.21...@times-components/article-extras@0.15.22) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.15.21](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.20...@times-components/article-extras@0.15.21) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.15.20](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.15.19...@times-components/article-extras@0.15.20) (2022-07-11)
 
 **Note:** Version bump only for package @times-components/article-extras