+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.10.5](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.10.4...@times-components/article-image@7.10.5) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
+## [7.10.4](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.10.3...@times-components/article-image@7.10.4) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
+## [7.10.3](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.10.2...@times-components/article-image@7.10.3) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
+## [7.10.2](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.10.1...@times-components/article-image@7.10.2) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
 ## [7.10.1](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.10.0...@times-components/article-image@7.10.1) (2022-07-06)
 
 **Note:** Version bump only for package @times-components/article-image