+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.18](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.8.17...@times-components/article-label@2.8.18) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
+## [2.8.17](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.8.16...@times-components/article-label@2.8.17) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
+## [2.8.16](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.8.15...@times-components/article-label@2.8.16) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
+## [2.8.15](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.8.14...@times-components/article-label@2.8.15) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
 ## [2.8.14](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.8.13...@times-components/article-label@2.8.14) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/article-label