+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.85.11](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.10...@times-components/article-skeleton@1.85.11) (2022-07-13)
+
+
+### Bug Fixes
+
+* **TDP-398:** Added embedUrl ([#2803](https://github.com/newsuk/times-components/issues/2803)) ([60b9c1d](https://github.com/newsuk/times-components/commit/60b9c1de5e5c5f0e00dec87f9672eca6e73b9ec8))
+
+
+
+
+
+## [1.85.10](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.9...@times-components/article-skeleton@1.85.10) (2022-07-13)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.85.9](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.8...@times-components/article-skeleton@1.85.9) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.85.8](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.7...@times-components/article-skeleton@1.85.8) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.85.7](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.6...@times-components/article-skeleton@1.85.7) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.85.6](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.5...@times-components/article-skeleton@1.85.6) (2022-07-11)
+
+
+### Bug Fixes
+
+* **TDP-1704:** Remove react-native from jest-serializer package ([#3015](https://github.com/newsuk/times-components/issues/3015)) ([5b6da8f](https://github.com/newsuk/times-components/commit/5b6da8f1392c3b42a9c0944d44be3d9ed3e7db2f))
+
+
+
+
+
 ## [1.85.5](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.85.4...@times-components/article-skeleton@1.85.5) (2022-07-11)
 
 **Note:** Version bump only for package @times-components/article-skeleton