+++ packages/button/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.27](https://github.com/newsuk/times-components/compare/@times-components/button@2.5.26...@times-components/button@2.5.27) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/button
+
+
+
+
+
+## [2.5.26](https://github.com/newsuk/times-components/compare/@times-components/button@2.5.25...@times-components/button@2.5.26) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/button
+
+
+
+
+
+## [2.5.25](https://github.com/newsuk/times-components/compare/@times-components/button@2.5.24...@times-components/button@2.5.25) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/button
+
+
+
+
+
+## [2.5.24](https://github.com/newsuk/times-components/compare/@times-components/button@2.5.23...@times-components/button@2.5.24) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/button
+
+
+
+
+
 ## [2.5.23](https://github.com/newsuk/times-components/compare/@times-components/button@2.5.22...@times-components/button@2.5.23) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/button