+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.11.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.11.4...@times-components/related-articles@6.11.5) (2022-07-13)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.11.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.11.3...@times-components/related-articles@6.11.4) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.11.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.11.2...@times-components/related-articles@6.11.3) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.11.2](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.11.1...@times-components/related-articles@6.11.2) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.11.1](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.11.0...@times-components/related-articles@6.11.1) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 # [6.11.0](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.10.7...@times-components/related-articles@6.11.0) (2022-07-11)