+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.2](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.10.1...@times-components/tracking@2.10.2) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
+## [2.10.1](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.10.0...@times-components/tracking@2.10.1) (2022-07-12)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
+# [2.10.0](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.9.32...@times-components/tracking@2.10.0) (2022-07-11)
+
+
+### Features
+
+* **TDP-1696:** add TcFlatlist component, refactor slice-layout ([#3022](https://github.com/newsuk/times-components/issues/3022)) ([32dc942](https://github.com/newsuk/times-components/commit/32dc942c7e0895453d7c80d2e30b6bcf466c9bdf))
+
+
+
+
+
+## [2.9.32](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.9.31...@times-components/tracking@2.9.32) (2022-07-11)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
 ## [2.9.31](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.9.30...@times-components/tracking@2.9.31) (2022-05-16)
 
 **Note:** Version bump only for package @times-components/tracking